ljh版本只完成了增添和筛选
