+++ src/views/courseCategory/index.vue
@@ -5,7 +5,7 @@
       <div class="filter-container">
         <div class="filter-item-container">
           <span>分类项数量</span>
-          <el-select v-model="form.search_categoryNum" placeholder="请选择">
+          <el-select v-model="form.search_categoryNum" clearable placeholder="请选择">
             <el-option v-for="item in options" :key="item.value" :label="item.label" :value="item.value">
             </el-option>
           </el-select>
@@ -13,14 +13,15 @@
         <div class="filter-item-container">
           <span>创建时间</span>
           <el-date-picker v-model="value_data" type="datetimerange" range-separator="至" start-placeholder="开始日期"
-            end-placeholder="结束日期">
+            end-placeholder="结束日期"
+            value-format="yyyy-MM-dd HH:mm:ss">
           </el-date-picker>
         </div>
         <div class="filter-item-container search">
           <el-input v-model="form.search_categoryKey" placeholder="请输入课程分类名的关键字"></el-input>
         </div>
 
-        <el-button type="primary" icon="el-icon-search">筛选</el-button>
+        <el-button type="primary" icon="el-icon-search" @click="fetchData">筛选</el-button>
       </div>
       <div class="filter-container">
         <span>
@@ -36,7 +37,7 @@
               </div>
             </el-col>
           </el-row>
-<!-- 添加分类项的弹窗 -->
+
           <el-dialog custom-class="scroll" width=90% title="课程分类管理/添加分类" :visible.sync="dialogFormVisible">
             <div class="height_action">请填写课程分类相关信息</div>
             <div class="demo-input-suffix">
@@ -154,126 +155,6 @@
               </el-row>
             </div>
           </el-dialog>
-<!-- 修改分类项弹窗 -->
- <el-dialog custom-class="scroll" width=90% title="课程分类管理/编辑分类" :visible.sync="dialogTableVisible">
-            <div class="height_action">请填写课程分类相关信息</div>
-            <div class="demo-input-suffix">
-<!--              <el-row :gutter="10" pading="30px">
-                <el-col :span="3">
-                  <h3>创建人:</h3>
-                </el-col>
-                <el-col :span="7">
-                  <el-input placeholder="请输入名字" v-model="action_name_input"></el-input>
-                </el-col>
-              </el-row> -->
-              <el-row :gutter="10">
-                <el-col :span="3">
-                  <h3>课程分类组名称:</h3>
-                </el-col>
-                <el-col :span="7">
-                  <el-input  v-model=" change_course_class.className"></el-input>
-                </el-col>
-              </el-row>
-              <el-row :gutter="10">
-                <el-col :span="3">
-                  <h3>课程分类组封面:</h3>
-                </el-col>
-                <el-col :span="10">
-                  <el-upload
-                  auto-upload=false
-                  limit=1
-                  action="http://106.55.25.94:8080/api/user/modifyHptIos"
-                  :data="transformPhoto"
-                  accept=".jpg,.png"
-                  drag="true"
-                  list-type="picture-card"
-                  :on-preview="handlePictureCardPreview"
-                  :on-remove="handleRemove"
-                  :on-success="change_imgs" 
-                  :file-list="show_img"
-                  name="headPortrait"
-                  v-model="transformPhoto.file"
-                  >
-                    <i class="el-icon-plus"></i>
-                    <div class="el-upload__tip" slot="tip">只能上传jpg/png文件，且最多仅支持上传一张图片</div>
-                    <div class="el-upload__tip" slot="tip"> </div>
-                  </el-upload>
-                </el-col>
-              </el-row>
-              <div class="height_action"></div>
-              <el-row :gutter="10" pading="30px">
-                <el-col :span="5" :offset="5">
-                  <el-button type="success" @click="updata_course_class();">确认更新分类组</el-button>
-                </el-col>
-                <el-col :span="5">
-                  <el-button @click="dialogTableVisible = false">取消</el-button>
-                </el-col>
-              </el-row>
-
-              <div class="height_action"></div>
-              <el-divider><font color="red">添加分类项前请务必先添加分类组</font></el-divider>
-
-              <div class="height_action"></div>
-              <el-row :gutter="10">
-                <el-col :span="3">
-                  <h3>添加分类项:</h3>
-                </el-col>
-                <el-col :span="7">
-                  <el-button type="success" @click="change_change_Child">点击添加分类项</el-button>
-                </el-col>
-              </el-row>
-              <section>
-                <div class="height_action"></div>
-                <template v-for="(item,index) in  change_course_class.courseClassList">
-                  <span>
-                    <el-row :gutter="20">
-                      <el-col :span="3">
-                        <h3>分类项名{{index + 1}}</h3>
-                      </el-col>
-                      
-                      
-                      <el-col :span="7">
-                        <el-input  v-model="change_course_class.courseClassList[index].classValue"></el-input>
-                      </el-col>
-                      <el-col :span="2">
-                        <el-button @click="delect_change_action(index)">取消</el-button>
-                      </el-col>
-                    </el-row>
-                    <el-row :gutter="10">
-                      <el-col :span="3">
-                        <h3>分类项名{{index + 1}}封面:</h3>
-                      </el-col>
-                      <el-col :span="10">
-                        <el-upload auto-upload=false limit=1  action="http://106.55.25.94:8080/api/user/modifyHptIos"
-                          :data="transformPhoto"
-                          :before-upload="change_number(index)"
-                          :on-success="change_photo"
-                          :file-list="change_courseClassList[index]"
-                          name="headPortrait"
-                          list-type="picture-card" :on-preview="handlePictureCardPreview" :on-remove="handleRemove">
-                          <i class="el-icon-plus"></i>
-                          <div class="el-upload__tip" slot="tip">只能上传jpg/png文件，且不超过500kb</div>
-                          <div class="el-upload__tip" slot="tip"> </div>
-
-                        </el-upload>
-                      </el-col>
-                    </el-row>
-                    <div class="height_action"></div>
-                  </span>
-                </template>
-              </section>
-
-              <div class="height_action"></div>
-              <el-row :gutter="10" pading="30px">
-                <el-col :span="5" :offset="5">
-                  <el-button type="success" @click="updata_course_class();TableVisible();">确认添加分类项</el-button>
-                </el-col>
-                <el-col :span="5">
-                  <el-button @click="dialogTableVisible = false">取消</el-button>
-                </el-col>
-              </el-row>
-            </div>
-          </el-dialog>
 
           <el-table v-loading="listLoading" :data="list" border style="width: 100%">
             <el-table-column fixed prop="courseClassId" label="课程编号" width="100">
@@ -295,8 +176,8 @@
 
             <el-table-column label="操作" width="180">
               <template slot-scope="scope">
-                <el-button size="mini" @click="dialogTableVisible = true;handleEdit(scope.row);show_imgs()">编辑</el-button>
-                <el-button size="mini" type="danger" @click="handleDelete(scope.row.courseClassId)">删除</el-button>
+                <el-button size="mini" @click="handleEdit(scope.$index)">编辑</el-button>
+                <el-button size="mini" type="danger" @click="handleDelete(scope.$index)">删除</el-button>
               </template>
             </el-table-column>
 
@@ -360,10 +241,8 @@
   import {
     getList,
     addCourseList,
-    deleteCourseClass,
     addCourseGroupList,
-    updataCourseClass,
-    modifyHptIos,
+    modifyHptIos
   } from '@/api/courseClass'
   import { getToken } from '@/utils/auth'
   
@@ -374,7 +253,6 @@
     data() {
       return {
         dialogFormVisible: false,
-        dialogTableVisible: false,
         listLoading: true,
         options: [{
             value: 1,
@@ -418,7 +296,6 @@
           }
         ],
         child: 0,
-        chang_child:0,
         list: [],
         courseClassList:[], //分类项列表
         form: {
@@ -440,17 +317,8 @@
           file:{},
           token:getToken()
         },
-        show_img:[{name:'hhh',url:''}],
-        
-        change_course_class:{
-           courseClassId:0,
-           className:'',
-           classUrl:'',
-           courseClassList:[]
-           },  
-        courseClass_number:0,//记录正在操作的分类项
-        change_courseClass_number:0,
-        change_courseClassList:[]
+        courseClass_number:0//记录正在操作的分类项
+        ,value_data:[]
       }
     },
     created() {
@@ -459,7 +327,7 @@
     methods: {
       fetchData() {
         this.listLoading = true
-        getList().then(response => {
+        getList(this.form.search_categoryNum,this.value_data[0],this.value_data[1],this.form.search_categoryKey).then(response => {
           this.list = response.data.adminCourseClassDtoList
           this.listLoading = false
         })
@@ -473,7 +341,6 @@
           cancelButtonText: '取消',
           type: 'warning'
         }).then(() => {
-           this.delete_course_class(index)
           this.$message({
             type: 'success',
             message: '删除成功!'
@@ -509,10 +376,7 @@
       },
       //照片变成Url
       insert_photo(response){
-        this.courseClassList[this.courseClass_number].classUrl=response.data;
-      },
-      change_photo(response){
-        this.change_course_class.courseClassList[this.change_courseClass_number].classUrl=response.data;
+        this.courseClassList[this.courseClass_number].classUrl=response.data
       },
       //照片存储
       transformPhoto2List(response){
@@ -535,75 +399,6 @@
       //删除list
       delect_action(val){
         this.courseClassList.splice(val,1)
-      },
-      delete_course_class(index){
-        deleteCourseClass(getToken(),index)
-      },
-      //修改分类项
-       changeChild() {
-        this.change_child++
-        this.change_course_class.courseClassList.push({
-          type_old: 'child',
-          token:getToken(),
-          className:"",
-          classValue:"",
-          classUrl:''
-        })
-      },
-      change_change_Child() {
-        this.child++
-        this.change_course_class.courseClassList.push({
-          type_old: 'child',
-          token:getToken(),
-          className:"",
-          classValue:"",
-          classUrl:''
-        })
-      },
-      //修改课程信息
-       handleEdit(row){
-         console.log(row);
-         
-        this.change_course_class.className=row.className;
-        this.change_course_class.classUrl=row.classUrl;
-        this.change_course_class.courseClassId=row.courseClassId;
-        this.change_course_class.courseClassList=row.courseClassList;
-        this.change_child=row.courseClassList.length;
-        for(let t1=0;t1<row.courseClassList.length;t1++){
-          this.change_courseClassList.push([{
-            name:row.courseClassList[t1].className,
-            url:row.courseClassList[t1].classUrl
-          }])
-        }
-        // this.change_course_class.courseClassList.classListUrl=row.classListUrl;
-
-      },
-       change_imgs(response){
-         this.change_course_class.classUrl=response.data
-      },
-      show_imgs(){
-       this.show_img[0].url=this.change_course_class.classUrl;
-       this.show_img[0].name=this.change_course_class.className;
-      },
-      updata_course_class(){ 
-        console.log(this.change_course_class);
-       updataCourseClass(this.change_course_class)
-        .then(() => {
-          this.$message({
-            type: 'success',
-            message: '更新成功!'
-          });
-          
-        }).catch(() => {
-          this.$message({
-            type: 'info',
-            message: '更新失败'
-          });
-        });
-        
-      },
-      TableVisible(){
-        this.dialogTableVisible =false;
       }
     }
   }
